GPII-1596: Starting to write tests.
