+++ src/icons/gpii-icon@3x.png
