+++ src/icons/gpii-icon@6x.png
