+++ tests.js
@@ -4,5 +4,6 @@
 var app = require("electron").app;
 
 app.on("ready", function () {
-    require("./node_modules/gpii-windows/tests/UnitTests.js");
+    //require("./node_modules/gpii-windows/tests/UnitTests.js");
+    require("./tests/testMain.js");
 });
